+++ src/main/java/com/example/dto/product/ProductRequest.java
@@ -3,7 +3,6 @@ package com.example.dto.product;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
-import java.math.BigDecimal;
 
 @Data
 public class ProductRequest {
@@ -13,7 +12,7 @@ public class ProductRequest {
     private String description;
 
     @DecimalMin(value = "0.01", message = "Price must be greater than 0")
-    private BigDecimal price;
+    private float price;
 
     @NotBlank(message = "Category is required")
     private String category;
